Silicon Entropy Engine V0.7 - Phase 27F: Gated MoE + Expert Profiles
Completes Phase 27E/F: Gated Expert Eligibility is now a first-class
architectural primitive, and expert sets are separated from speed profiles.

Core architecture changes:
- moe_engine: add sum_w_elig[] / n_elig_steps[] telemetry; moe_update_gated
  accumulates eligible-only averages; moe_avg_weight_elig() accessor added
- see_codec: add SEE_FLAG_TOK_PREV_ELIG (0x20u) to archive header — fixes
  encode/decode SHA-256 symmetry bug (eligibility was lost on decode)
- see_codec: add n_elig_bytes, tokprev_alnum_start_bpb, n_alnum_start_bytes
  telemetry to SeeAuditResult; see_audit_result_print shows eligibility block

Expert profile system:
- SeeCodecConfig gains expert_profile + speed_profile fields
- apply_expert_profile(): general = LZ6+TOKPFX, text = LZ6+TOKPFX+TOK_PREV_ELIG
- CLI: --expert-profile general|text|experimental, --speed full|accurate|fast
  (--profile kept as alias for backward compat)

Phase 27F audit results (natural_text):
- text profile gain: -0.0445 BPB vs general (-0.0023 c_code, +0.0032 markdown)
- W_TOKPREV when-eligible: 0.8030 (22.9% eligible bytes = ALNUM_START bytes)
- TOKPREV BPB@START: 7.09 vs global 7.79 (+0.70 gap — gain concentrated at starts)
- SHA-256: PASS on all 3 datasets after flag fix

+++ src/lz_topn.h
@@ -18,8 +18,12 @@ typedef struct {
     uint16_t count;
 } LzTopSlot;
 
+// Key is always uint64_t; callers mask to 32/48/64 bits before passing.
+// LZ4 → key = lz_ctx & 0xFFFFFFFF
+// LZ6 → key = lz_ctx & 0xFFFFFFFFFFFFULL
+// LZ8 → key = lz_ctx  (full 64 bits)
 typedef struct {
-    uint32_t  key;
+    uint64_t  key;
     uint32_t  total;
     LzTopSlot top[LZ_TOP_N];
 } LzEntry;
@@ -31,12 +35,12 @@ LzEntry* lz_table_alloc(void);
 void lz_table_free(LzEntry* table);
 
 // Look up (or claim an empty slot for) the given key.
-LzEntry* lz_lookup(LzEntry* table, uint32_t key);
+LzEntry* lz_lookup(LzEntry* table, uint64_t key);
 
 // Fill p_out[256] with the smoothed LZ probability distribution for entry e.
 void lz_build_probs(const LzEntry* e, float lz_K, float* p_out);
 
 // Record that byte `target` was observed under context key `ctx_key`.
-void lz_update(LzEntry* e, uint32_t ctx_key, uint8_t target);
+void lz_update(LzEntry* e, uint64_t ctx_key, uint8_t target);
 
 #endif // LZ_TOPN_H
